refactor: authentication_service wrong naming convetion
